move the rest of the learn content and update nav and footer
